Usage

The Option and Result type.

Err{E} and Some{T} are simple wrapper structs with little functionality in themselves. Some is from Base Julia, Err from this package. They are only used to wrap a value to represent a correctly generated result (Some) or an error (Err). They are two distinct types, such that a correctly-produced value of type T can be distinguished from an error of type T (e.g. an integer result from an integer error code).

Result{T, E} is merely an alias for Union{Some{T}, Err{E}}. Thus, the result of a failable funtion that can give a result of type T or an error of type E is a Result{T, E}. Do remember Julia's parametric invariance: It is NOT true that Some([1]) isa Result{Vector, String}, whereas it IS true that it is a Result{<:Vector, String}.

In many cases, the error type does not need to be specified because there is only one relevant error condition. In these situations, it's easier to use Option{T}. This is simply an alias of Union{Some{T}, Nothing}.

Basic usage

Always typeassert any function that returns an error type. The whole point of ErrorTypes is to encode error states in return types, and be specific about these error states. While ErrorTypes will technically work fine without function annotations, I highly recommend annotating return types:

Do this:

invert(x::Integer)::Option{Float64} = iszero(x) ? nothing : Some(1/x)

And not this:

invert(x::Integer) = iszero(x) ? nothing : Some(1/x)

The typeassert will cause your functions to fail immediately in testing, and will also be caught by static type checkers such as JET.jl. Note that typeasserts applied to functions are different than other typeasserts, which will not be caught by JET.jl.

Convenience functions

The entire point of this package is static types made convenient. Since unions are already part of Base Julia, the idea of this package is to provide convenience functions. Particularly useful functions are something_or, and_then, and flatten.

Conversion

Converting to and from Err is intentionally limited to prevent silent errors. Only two kinds of conversions are possible:

Converting an Err to Option, which results in nothing, and

Converting an Err{E1} to an Err{>:E1}, which simply re-wraps the input value.

@?

If you make an entire codebase of functions returning Results, it can get bothersome to constantly check if function calls contain error values and propagate those error values. To make this process easier, use the macro @?, which automatically propagates any error values. If this is applied to some expression x evaluating to a Result or Option containing an error value, the macro will evaluate to return x. If it contains a result value, the macro is evaluated to something(x).

For example, suppose you want to implement a safe version of the harmonic mean function, which in turn uses a safe version of div:

safe_div(a::Integer, b::Real)::Option{Float64} = iszero(b) ? nothing : Some(a/b)

function harmonic_mean(v::AbstractArray{<:Integer})::Option{Float64}
    sm = 0.0
    for i in v
        invi = safe_div(1, i)
        isnothing(invi) && return nothing
        sm += something(invi)
    end
    return safe_div(length(v), sm)
end

In this function, we constantly have to check whether safe_div returned the error value, and return that from the outer function in that case. That can be more concisely written as:

function harmonic_mean(v::AbstractArray{<:Integer})::Option{Float64}
    sm = 0.0
    for i in v
        sm += @? safe_div(1, i)
    end
    return safe_div(length(v), sm)
end

When to use an error type vs throw an error

The error handling mechanism provided by ErrorTypes is a distinct method from throwing and catching errors. None is superior to the other in all circumstances.

The handling provided by ErrorTypes is faster, safer, and more explicit. For most functions, you can use ErrorTypes. However, you can't only rely on it. Imagine a function A returning Option{T1}. A is called from function B, which can itself fail and returns an Option{T2}. However, now there are two distinct error states: Failure in A and failure in B. So what should B return? Result{T2, Enum{E1, E2}}, for some Enum type? But then, what about functions calling B? Where does it end?

In general, it's un-idiomatic to "accumulate" error states like this. You should handle an error state when it appears, and usually not return it far back the call chain.

More importantly, you should distinguish between recoverable and unrecoverable error states. The unrecoverable are unexpected, and reveals that the program went wrong somehow. If the program went somewhere it shouldn't be, it's best to abort the program and show the stack trace, so you can debug it - here, an ordinary exception is better. If the errors are known to be possible beforehand, using ErrorTypes is better. For example, a program may use exceptions when encountering errors when parsing "internal" machine-generated files, which are supposed to be of a certain format, and use error types when parsing user input, which must always be expected to be possibly fallible.

Because error types are so easily converted to exceptions (using unwrap and expect), internal library functions should preferably use error types.

Reference

ErrorTypes.Option — Type

Option{T}

Alias for Union{Some{T}, Nothing}.

ErrorTypes.Result — Type

Result{T, E}

Alias for Union{Some{T}, Err{E}}.

ErrorTypes.Err — Type

Err{E}

Wrapper type for error results of type E. Used in Union{Some{T}, Err{E}}.

ErrorTypes.and_then — Method

and_then(f, ::Type{T}, x::Union{Option{T}, Result{T}})

If x is something, return Some(f(something(x)))::T, else return x.

WARNING If f(something(x)) is not a T, this functions throws an error.

ErrorTypes.expect — Method

expect(x::Union{Option, Result}, s::AbstractString)

If x is of the associated error type, error with message s. Else, call something on the value.

ErrorTypes.expect_err — Method

expect_err(x::Union{Result, Option}, s::AbstractString)

If x is an Err or nothing, return the wrapped value. Else, throw an error with message s.

ErrorTypes.flatten — Method

flatten(x::Option{Option{T}})

Convert an Option{Option{T}} to an Option{T}.

Examples

julia> flatten(Some(Some("x")))
Some("x")

julia> flatten(Some(nothing))

ErrorTypes.something_or — Method

something_or(x::Result, v)

If x is an error value, return v. Else, unwrap x and return its content.

ErrorTypes.unwrap — Method

unwrap(x::Union{Option, Result})

Like expect, but throws a generic TypeError.

ErrorTypes.unwrap_err — Method

unwrap_err(x::Union{Option, Result})

Like expect_err, but throws a generic TypeError.

ErrorTypes.@? — Macro

@?(expr)

Propagate a Result or Option with error value to the outer function. Evaluate expr, which should return a Union{Option, Result} x. If x is a Some, evaluate to something(x). Else, evaluates to return x.

Example

julia> (f(x::Option{T})::typeof(x)) where T = Some(@?(x) + one(T));

julia> f(Some(1.0)), f(nothing)
(Some(2.0), nothing)

ErrorTypes.@something_or — Macro

@something_or(expr, exec)

Evaluate expr to a Result or Option. If expr is a error value, evaluates to exec. Else, evaluate something(expr).

Examples

julia> safe_inv(x)::Option{Float64} = iszero(x) ? nothing : Some(1/x);

julia> function skip_inv_sum(it)
    sum = 0.0
    for i in it
        sum += @something_or safe_inv(i) continue
    end
    sum
end;

julia> skip_inv_sum([2,1,0,1,2])
3.0