Usage

Types and how to construct them

Option{T} encodes either the presence of a T or the absence of one. You should use this type to encode either the successful creation of T, or a failure that does not need any information to explain itself. Like its sibling Result, Option is a sum type from the package SumTypes - think of it like a fancy Enum.

An Option contains one of two variants, None and Thing. Like other sum types, you should usually not construct an Option directly. Instead, the constructors for the types None and Thing creates an Option, wrapping them - e.g. None{Int}() creates an Option{Int} containing a None. Similarly, you should not construct a None or Thing directly - these types are only useful as the content of an Option.

You should construct an Option{T} wrapping a Thing with Thing(::T).

You can construct an Option{T} warpping a None object with None{Int}(). Normally, however, you should use the object none (singleton of None{Nothing}) and convert it to an Option{T}. See how in the section "Basic usage" below.

Result{O, E} has the two variants Ok, representing the successful creation of an O, or else an Err, representing some object of type E, carrying information about the error state. You should use Result instead of Option when the error result needs to carry information.

You can construct Result objects by Ok{O, E}(::O) or Err{O, E}(::E). However, normally, you should instead use the simpler constructors Ok(::O) and Err(::E). These constructors creates ResultConstructor objects, which can be converted to the correct Result types.

Basic usage

Always typeassert any function that returns an error type. The whole point of ErrorTypes is to encode error states in return types, and be specific about these error states. While ErrorTypes will technically work fine without function annotations, I highly recommend annotating return types:

Do this (note none is automatically converted):

invert(x::Integer)::Option{Float64} = iszero(x) ? none : Thing(1/x)

And not this:

invert(x::Integer) = iszero(x) ? None{Float64}() : Thing(1/x)

Also, by annotating return functions, you can use the simpler constructors of error types:

You can use none, which is automatically converted to Option containing a None.

You can use Err(x) and Ok(x) to create ResultConstructors, which are converted by the typeassert to the correct Result.

Conversion

Apart from the two special cases above: The conversion of none (of type None) to Option, and Err(x) and Ok(x) (of type ResultConstructor) to Result, error types can only convert to each other in certain circumstances. This is intentional, because type conversion is a major source of mistakes.

An object of type Option and Result can be converted to its own type, i.e. convert(T, ::T) works.

An Option{T} can be converted to an Option{T2}, if T <: T2. You intentionally cannot convert e.g. an Option{Int} to an Option{UInt}.

A Result{O, E} can be converted to a Result{O2, E2} if O <: O2 and E <: E2.

It is intentionally NOT possible to e.g. convert a Result{Int, String} containing an Ok to a Result{Int, Int}, even if the Ok value contains an Int which is allowed in both of the Result types. The reason for this is that if it was allowed, whether or not conversions threw errors would depend on the value of an error type, not the type. This type-unstable behaviour would defeat idea behind this package, namely to present edge cases as types, not values.

@?

If you make an entire codebase of functions returning Results and Options, it can get bothersome to constantly check if function calls contain error values and propagate those error values. To make this process easier, use the macro @?, which automatically propagates any error values. If this is applied to some expression x evaluating to a Result or Option containing an error value, the macro will evaluate to something equivalent to return Err(unwrap_err(x)) or return none, respectively. If it contains a result value, the macro is evaluated to the equivalent of unwrap(x).

For example, suppose you want to implement a safe version of the harmonic mean function, which in turn uses a safe version of div:

safe_div(a::Integer, b::Real)::Option{Float64} = iszero(b) ? none : Thing(a/b)

function harmonic_mean(v::AbstractArray{<:Integer})::Option{Float64}
    sm = 0.0
    for i in v
        invi = safe_div(1, i)
        is_none(invi) && return none
        sm += unwrap(invi)
    end
    res = safe_div(length(v), sm)
    is_none(res) && return none
    return Thing(unwrap(res))
end

In this function, we constantly have to check whether safe_div returned the error value, and return that from the outer function in that case. That can be more concisely written as:

function harmonic_mean(v::AbstractArray{<:Integer})::Option{Float64}
    sm = 0.0
    for i in v
        sm += @? safe_div(1, i)
    end
    Thing(@? safe_div(length(v), sm))
end

When to use an error type vs throw an error

The error handling mechanism provided by ErrorTypes is a distinct method from throwing and catching errors. None is superior to the other in all circumstances.

The handling provided by ErrorTypes is faster, safer, and more explicit. For most functions, you can use ErrorTypes. However, you can't only rely on it. Imagine a function A returning Option{T1}. A is called from function B, which can itself fail and returns an Option{T2}. However, now there are two distinct error states: Failure in A and failure in B. So what should B return? Result{T2, Enum{E1, E2}}, for some Enum type? But then, what about functions calling B? Where does it end?

In general, it's un-idiomatic to "accumulate" error states like this. You should handle an error state when it appears, and usually not return it far back the call chain.

More importantly, you should distinguish between recoverable and unrecoverable error states. The unrecoverable are unexpected, and reveals that the program went wrong somehow. If the program went somewhere it shouldn't be, it's best to abort the program and show the stack trace, so you can debug it - here, an ordinary exception is better. If the errors are known to be possible beforehand, using ErrorTypes is better. For example, a program may use exceptions when encountering errors when parsing "internal" machine-generated files, which are supposed to be of a certain format, and use error types when parsing user input, which must always be expected to be possibly fallible.

Because error types are so easily converted to exceptions (using unwrap and expect), internal library functions should preferably use error types.

Reference

ErrorTypes.Err — Method

Err

The error state of a Result{O, E}, carrying an object of type E. For convenience, Err(x) creates a dummy value that can be converted to the appropriate Result type.

ErrorTypes.Ok — Method

Ok

The success state of a Result{O, E}, carrying an object of type O. For convenience, Ok(x) creates a dummy value that can be converted to the appropriate Result type.

ErrorTypes.Option — Type

Option{T}

Alias for Result{T, Nothing}

ErrorTypes.Result — Type

Result{O, E}

A sum type of either Ok{O} or Err{E}. Used as return value of functions that can error with an informative error object of type E.

ErrorTypes.and_then — Method

and_then(f, ::Type{T}, x::Result{O, E})

If is a result value, apply f to unwrap(x), else return the error value. Always returns an Option{T} / Result{T, E}.

WARNING If f(unwrap(x)) is not a T, this functions throws an error.

ErrorTypes.expect — Function

expect(x::Result, s::AbstractString)

If x is of the associated error type, error with message s. Else, return the contained result type.

ErrorTypes.expect_err — Method

expect_err(x::Result, s::AbstractString)

If x contains an Err, return the content of the Err. Else, throw an error with message s.

ErrorTypes.flatten — Method

flatten(x::Option{Option{T}})

Convert an Option{Option{T}} to an Option{T}.

Examples

julia> flatten(some(some("x")))
Option{String}: Ok("x")

julia> flatten(some(none(Int)))
Option{Int64}: Err(nothing)

ErrorTypes.unwrap — Function

unwrap(x::Result)

If x is of the associated error type, throw an error. Else, return the contained result type.

ErrorTypes.unwrap_err — Method

unwrap_err(x::Result)

If x contains an Err, return the content of the Err. Else, throw an error.

ErrorTypes.@? — Macro

@?(expr)

Propagate a Result with Err value to the outer function. Evaluate expr, which should return a Result. If it contains an Ok value x, evaluate to the unwrapped value x. Else, evaluates to return Err(x)t.

Example

julia> (f(x::Option{T})::Option{T}) where T = Ok(@?(x) + one(T));

julia> f(some(1.0)), f(none(Int))
(Option{Float64}: Ok(2.0), Option{Int64}: Err(nothing))

ErrorTypes.@unwrap_or — Macro

@unwrap_or(expr, exec)

Evaluate expr to a Result. If expr is a error value, evaluate exec and return that. Else, return the wrapped value in expr.

Examples

julia> safe_inv(x)::Option{Float64} = iszero(x) ? none : Ok(1/x);

julia> function skip_inv_sum(it)
    sum = 0.0
    for i in it
        sum += @unwrap_or safe_inv(i) continue
    end
    sum
end;

julia> skip_inv_sum([2,1,0,1,2])
3.0