Usage

Types and how to construct them

Option{T} encodes either the presence of a T or the absence of one. You should use this type to encode either the successful creation of T, or a failure that does not need any information to explain itself. Like its sibling Result, Option is a sum type from the pacage SumTypes.

An Option contains two variants, None and Thing. Like other sum types, you should usually not construct an Option directly. Instead, the constructors for the types None and Thing creates an Option wrapping them. Similarly, you should not construct a None or Thing directly - these types are only useful as the content of an Option.

You should construct an Option{T} wrapping a Thing with Thing(::T).

You can construct an Option{T} warpping a None object with ErrorTypes.None{Int}(). Normally, however, you should use the object none (singleton of None{Nothing}) and convert it to an Option{T}. See how in the section "Basic usage".

Result{O, E} has the two variants Ok, representing the successful creation of an O, or else an E, representing some object carrying information about the error. You should use Result instead of Option when the error needs to be explained.

You can construct Result objects by Ok{O, E}(::O) or Err{O, E}(::Err). However, normally, you can instead use the simpler constructors Ok(::O) and Err(::E). These constructors creates ResultConstructor objects, which can be converted to the correct Result types.

Basic usage

Always typeassert any function that returns an error type. The whole point of ErrorTypes is to encode error states in return types, and be specific about these error states. While ErrorTypes will technically work fine without function annotation, I highly recommend annotating return types:

Yes (note none is automatically converted):

invert(x::Integer)::Option{Float64} = iszero(x) ? none : Thing(1/x)

No:

invert(x::Integer) = iszero(x) ? ErrorTypes.None{Float64}() : Thing(1/x)

Also, by annotating return functions, you can use the simpler constructors of error types:

You can use none, which is automatically converted to Option containing a None.

You can use Err(x) and Ok(x) to create ResultConstructors, which are converted by the typeassert to the correct Result.

Conversion

Apart from the two special cases above: The conversion of none (of type None) to Option, and Err(x) and Ok(x) (of type ResultConstructor) to Result, error types can only convert to each other in certain circumstances. This is intentional, because type conversions is a major source of mistakes.

An object of type Option and Result can be converted to its own type.

An Option{T} containing a Thing can be converted to an Option{T2}, if T <: T2. You intentionally cannot convert e.g. an Option{Int} to an Option{UInt}.

A Result{O, E} can be converted to a Result{O2, E2} if O <: O2 and E <: E2.

It is intentionally NOT possible to e.g. convert a Result{Int, String} containing an Ok to a Result{Int, Int}, even if the Ok value contains an Int which is allowed in both of the Result types. The reason for this is that if it was allowed, whether or not conversions threw errors would depend on the value of an error type, not the type. The idea behind this package is to present edge cases as types, not values.

@?

If you make an entire codebase of functions returning Results and Options, it can get bothersome to constantly check if function calls contain error values and propagate those error values. To make this process easier, use the macro @?, which automatically propagates any error values. If this is applied to some expression x evaluating to a Result or Option containing an error value, the macro will evaluate to something equivalent to return Err(unwrap_err(x)) or return none, respectively. If it contains a result value, the macro is evaluated to the equivalent of unwrap(x).

For example, suppose you want to implement a safe version of the harmonic mean function, which in turn uses a safe version of div:

safe_div(a::Integer, b::Float64)::Option{Float64} = iszero(b) ? none : Thing(a/b)

function harmonic_mean(v::AbstractArray{<:Integer})::Option{Float64}
    sm = 0.0
    for i in v
        invi = safe_div(1, i)
        is_none(invi) && return none
        sm += unwrap(invi)
    end
    res = safe_div(length(v), sm)
    is_none(res) && return none
    return Thing(unwrap(res))
end

In this function, we constantly have to check whether safe_div returned the error value, and return that from the outer function in that case. That can be more concisely written as:

function harmonic_mean(v::AbstractArray{<:Integer})::Option{Float64}
    sm = 0.0
    for i in v
        sm += @? safe_div(1, i)
    end
    Thing(@? safe_div(length(v), sm))
end

Reference

ErrorTypes.Err — Method

Err

The error state of a Result{O, E}, carrying an object of type E. For convenience, Err(x) creates a dummy value that can be converted to the appropriate Result type.

ErrorTypes.None — Type

None{T}

The variant of Option signifying absence of a value of type T. The singleton none is the instance of None{Nothing} (NOT an Option), and can be freely converted to any Option value containing a None.

ErrorTypes.Ok — Method

Ok

The success state of a Result{O, E}, carrying an object of type O. For convenience, Ok(x) creates a dummy value that can be converted to the appropriate Result type.

ErrorTypes.Option — Type

Option{T}

A sum type of either Thing{T} or None. Used instead of Result as return value of functions whose error state need not carry any information.

ErrorTypes.Result — Type

Result{O, E}

A sum type of either Ok{O} or Err{E}. Used as return value of functions that can error with an informative error object of type E.

ErrorTypes.Thing — Method

Thing(x::T)

Construct a Thing{T}(x), variant of Option signifying the presence of a value x.

ErrorTypes.and_then — Method

and_then(f, x::Union{Option, Result})

If x is an error value, return x. Else, apply f to the contained value and return it wrapped as the same type as x.

ErrorTypes.expect — Function

expect(x::Union{Result, Option}, s::AbstractString)

If x is of the associated error type, error with message s. Else, return the contained result type.

ErrorTypes.expect_err — Method

expect_err(x::Result, s::AbstractString)

If x contains an Err, return the content of the Err. Else, throw an error with message s.

ErrorTypes.expect_none — Method

expect_none(x::Option, s::AbstractString)

If x contains a None, return nothing. Else, throw an error with message s.

ErrorTypes.flatten — Method

flatten(x::Option{Option{T}})

Convert an Option{Option{T}} to an Option{T}.

Examples

julia> flatten(Thing(Thing("x")))
Option{String}: Thing("x")

julia> flatten(Thing(None{Int}()))
Option{Int64}: None()

ErrorTypes.unwrap — Function

unwrap(x::Union{Result, Option})

If x is of the associated error type, throw an error. Else, return the contained result type.

ErrorTypes.unwrap_err — Method

unwrap_err(x::Result)

If x contains an Err, return the content of the Err. Else, throw an error.

ErrorTypes.unwrap_none — Method

unwrap_none(x::Option)

If x contains a None, return nothing. Else, throw an error.

ErrorTypes.unwrap_or — Method

unwrap_or(x::Union{Option, Result}, v)

If x is an error value, return v. Else, unwrap x and return its content.

ErrorTypes.@? — Macro

@?(expr)

Propagate a Result or Option error value to the outer function. Evaluate expr, which should return a Result or Option. If it contains a result value x, evaluate to the unwrapped value x. Else, evaluates to return none for Option and return Err(x) for Result.

Example

julia> (f(x::Option{T})::Option{T}) where T = Thing(@?(x) + one(T))

julia> f(Thing(1.0)), f(None{Int}())
(Option{Float64}: Thing(2.0), Option{Int64}: None())

ErrorTypes.@unwrap_or — Macro

@unwrap_or(expr, exec)

Evaluate expr to a Result or Option. If expr is a error value, evaluate exec and return that. Else, return the wrapped value in expr.

Examples

julia> safe_inv(x)::Option{Float64} = iszero(x) ? none : Thing(1/x);

julia> function skip_inv_sum(it)
    sum = 0.0
    for i in it
        sum += @unwrap_or safe_inv(i) continue
    end
    sum
end;

julia> skip_inv_sum([2,1,0,1,2])
3.0